Add subscription URL environment logic
